re-copy skeleton

+++ skeleton-sp19/lab2/DebugPractice/DebugExercise1.java
@@ -4,9 +4,9 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        double quotient = (double)top / (double)bottom;
-        double result = Math.round(quotient);
-        return (int)result;
+        int quotient = top / bottom;
+        int result = Math.round(quotient);
+        return result;
     }
 
     public static void main(String[] args) {
